pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "제주 맛집 추천 챗봇"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.38.0",
    "google-generativeai>=0.8.0",
    "pandas>=2.2.0",
    "mcp>=1.13.1",
    "fastmcp>=2.11.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
]